Refactor server for better testing flow
