+++ package.json
@@ -24,7 +24,7 @@
     "eslint-plugin-react": "^7.4.0",
     "mocha": "^4.0.1",
     "nodemon": "^1.12.1",
-    "request": "^2.83.0"
+    "supertest": "^3.0.0"
   },
   "scripts": {
     "babel-node": "babel-node --ignore='/node_modules/'",
