+++ server/server.js
@@ -1,23 +1,6 @@
-import express from 'express';
-import graphqlHTTP from 'express-graphql';
-import { buildSchema } from 'graphql';
-import sequelize from '../database/database';
+import app from './app';
+import db from '../database/database';
 
-const app = express();
 const port = 3000;
-
-const schema = buildSchema(`
-  type Query {
-    hello: String
-  }
-`);
-
-const root = { hello: () => 'Hello world!' };
-
-app.use('/graphql', graphqlHTTP({
-  schema,
-  rootValue: root,
-  graphiql: true
-}));
 
 app.listen(port, () => console.log(`Listening on port ${port}`));
